ssbs: move tag parsing to its own module
